+++ R/biocExtract.R
@@ -14,12 +14,19 @@ NULL
 #' \code{RaggedExperiment} features. See
 #' \code{\link[RaggedExperiment]{RaggedExperiment-class}} for more details.
 #'
+#' @details A typical additional argument for this function passed down to
+#' lower level functions is the `names.field` which indicates the row names
+#' in the data. By default, it is the "Hugo_Symbol" column in the internal
+#' code that converts `data.frame`s to `SummarizedExperiment` representations
+#' (via the `.makeSummarizedExperimentFromDataFrame` internal function).
+#'
 #' @section type:
 #' Choices include:
 #' \itemize{
 #'     \item{clinical} - Get the clinical data slot
-#'     \item{RNASeqGene} - RNASeqGene
-#'     \item{RNASeq2GeneNorm} - Normalized
+#'     \item{RNASeqGene} - RNASeqGene - RNASeq v1
+#'     \item{RNASeqGene} - RNASeq2Gene - RNASeq v2
+#'     \item{RNASeq2GeneNorm} - RNASeq v2 Normalized
 #'     \item{miRNASeqGene} - micro RNA SeqGene
 #'     \item{CNASNP} - Copy Number Alteration
 #'     \item{CNVSNP} - Copy Number Variation
@@ -39,20 +46,36 @@ NULL
 #' @param object A \code{FirehoseData} object from which to extract data.
 #' @param type The type of data to extract from the "FirehoseData" object,
 #' see type section.
+#' @param ... Additional arguments passed to lower level functions that
+#' convert tabular data into Bioconductor object such as
+#' \code{.makeRangedSummarizedExperimentFromDataFrame} or
+#' \code{.makeRaggedExperimentFromDataFrame}
+#'
 #' @return Either an \linkS4class{SummarizedExperiment} object or a
 #' \linkS4class{RaggedExperiment} object.
 #'
+#' @md
+#'
 #' @author Marcel Ramos \email{marcel.ramos@@roswellpark.org}
 #'
-#' @examples \dontrun{
-#'     coadmut <- getFirehoseData("COAD", Mutation = TRUE)
-#'     biocExtract(coadmut, "Mutation")
-#' }
+#' @examples
+#'
+#' data(accmini)
+#' biocExtract(accmini, "RNASeq2Gene")
+#' biocExtract(accmini, "miRNASeqGene")
+#' biocExtract(accmini, "RNASeq2GeneNorm")
+#' biocExtract(accmini, "CNASNP")
+#' biocExtract(accmini, "CNVSNP")
+#' biocExtract(accmini, "Methylation")
+#' biocExtract(accmini, "Mutation")
+#' biocExtract(accmini, "RPPAArray")
+#' biocExtract(accmini, "GISTIC")
+#'
 #' @export biocExtract
 biocExtract <- function(object, type = c("clinical", "RNASeqGene",
-    "miRNASeqGene", "RNASeq2GeneNorm", "CNASNP", "CNVSNP", "CNASeq",
-    "CNACGH", "Methylation", "Mutation", "mRNAArray", "miRNAArray",
-    "RPPAArray", "GISTIC", "GISTICA", "GISTICT", "GISTICP")) {
+    "RNASeq2Gene", "miRNASeqGene", "RNASeq2GeneNorm", "CNASNP", "CNVSNP",
+    "CNASeq", "CNACGH", "Methylation", "Mutation", "mRNAArray", "miRNAArray",
+    "RPPAArray", "GISTIC", "GISTICA", "GISTICT", "GISTICP"), ...) {
     if (!identical(length(type), 1L))
         stop("Please specify a single data type")
     message("working on: ", type)
@@ -63,7 +86,7 @@ biocExtract <- function(object, type = c("clinical", "RNASeqGene",
         object <- .unNestList(object)
     }
     if (type == "clinical") { return(object) }
-    if (is(object, "matrix")) {
+    if (is.matrix(object)) {
         return(SummarizedExperiment(assays = SimpleList(object)))
     }
     if (is(object, "list") && !is(object, "DataFrame") &&
@@ -100,16 +123,7 @@ biocExtract <- function(object, type = c("clinical", "RNASeqGene",
 
     hasRanged <- .hasRangeNames(object)
     if (hasRanged) {
-        if (.hasInfo(object, "NCBI_Build")) {
-            GBuild <- .getBuild(object)
-        }
-        if (.hasConsistentRanges(object)) {
-            object <- .makeRangedSummarizedExperimentFromDataFrame(object,
-                build = if (exists("GBuild")) { GBuild } else { NA })
-        } else {
-            object <- .makeRaggedExperimentFromDataFrame(object,
-                build = if (exists("GBuild")) { GBuild } else { NA })
-        }
+        object <- .convertRangeBioc(object)
     } else {
         object <- .standardizeBC(object)
         metadat <- metadata(object)
@@ -117,4 +131,22 @@ biocExtract <- function(object, type = c("clinical", "RNASeqGene",
         metadata(object) <- metadat
     }
     return(object)
+}
+
+.convertRangeBioc <- function(object, ...) {
+    build <- unique(.getBuild(object))
+    build <- TCGAutils::correctBuild(build, "NCBI")
+    if (.hasConsistentRanges(object)) {
+        object <- .makeRangedSummarizedExperimentFromDataFrame(object,
+            build = build, ...)
+    } else {
+        split.field <- .findSampleCol(object)
+        if (is.na(split.field) || !length(split.field))
+            object <- .makeGRangesFromDataFrame(object,
+                build = build)
+        else
+            object <- .makeRaggedExperimentFromDataFrame(
+                object, split.field = split.field, build = build, ...)
+    }
+    object
 }